Further tuning of the Communications system. Now AbilityContours are considered a Source<Ability.Type> and Creatures were switched to be a Source<Skill>. Now both Ability Types and Skill Checks can be resolved from the Creature, which contains an AbilityContour.
In addition, JavaDocs were added to the mechanics.actions package. The RollCommand functionality was tweaked to adhere to the functionality in the test file, but still requires additional implementation as stated in the TODOs.

The Construct method was switched from returning the Builder object, to "deconstructing" the Construct to return a Constructor (Builder) with the same attributes as the original object.